Fix migration 177 if easuid table is not present

+++ migrations/versions/177_add_run_state_eas_folderstatus.py
@@ -15,10 +15,16 @@ import sqlalchemy as sa
 
 
 def upgrade():
-    op.add_column('easfoldersyncstatus', sa.Column('sync_should_run',
-                  sa.Boolean(), server_default=sa.sql.expression.true(),
-                  nullable=False))
+    from inbox.ignition import main_engine
+    engine = main_engine(pool_size=1, max_overflow=0)
+    if engine.has_table('easuid'):
+        op.add_column('easfoldersyncstatus', sa.Column('sync_should_run',
+                      sa.Boolean(), server_default=sa.sql.expression.true(),
+                      nullable=False))
 
 
 def downgrade():
-    op.drop_column('easfoldersyncstatus', 'sync_should_run')
+    from inbox.ignition import main_engine
+    engine = main_engine(pool_size=1, max_overflow=0)
+    if engine.has_table('easuid'):
+        op.drop_column('easfoldersyncstatus', 'sync_should_run')
